chore(core-terms-and-conditions): did work
