+++ packages/TermsAndConditions/FootnoteLink/FootnoteLink.md
@@ -1,0 +1,10 @@
+### Usage criteria
+
+```jsx
+<span>
+  Terms and conditions apply
+  <FootnoteLink number={1} />
+  <FootnoteLink number={2} />
+  <FootnoteLink number={3} />
+</span>
+```
